Updated tests to work with changes.
This may seem like cheating, however, previously the code escaped forward slashes, this is optional in JSON, and makes no difference on the Java or JavaScript ends (as far as I can tell).  The tests I changed were testing that the slash was being escaped, it wasn't, and didn't need to be.